// front/src/lib/types/api/v1/schema.d.ts
export interface paths {
// ...
  "/api/v1/members/logout": {
    /** 로그아웃 */
    post: operations["logout"];
  };
// ...
  "/api/v1/profileInventory/myInventory": {
    /** 플레이어 인벤토리 조회 */
    get: operations["getMyInventory"];
  };
// ...
  "/api/v1/members/test": {
    /** 테스트 */
    get: operations["test_1"];
  };
// ...
  "/api/v1/members/me": {
    /** 내 정보 */
    get: operations["getMe"];
  };
// ...
  "/api/v1/inventory/myInventory": {
    /** 플레이어 인벤토리 조회 */
    get: operations["getMyInventory_1"];
  };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
// ...
    ProfileInventoryDto: {
      /** Format: int64 */
      id: number;
      /** Format: date-time */
      createDate: string;
      /** Format: date-time */
      updateDate: string;
      profileIcon: components["schemas"]["ProfileDto"];
      isEquipped: boolean;
    };
// ...
    PlayerDto: {
      /** Format: int64 */
      id: number;
      /** Format: date-time */
      createDate: string;
      /** Format: date-time */
      modifyDate: string;
      nickname: string;
      /** Format: int32 */
      exp: number;
      /** Format: int32 */
      gems: number;
      /** Format: int32 */
      characterType: number;
      /** Format: int32 */
      backgroundVolume: number;
      /** Format: int32 */
      effectVolume: number;
      /** Format: int32 */
      editorAutoComplete: number;
      /** Format: int32 */
      editorAutoClose: number;
    };
// ...
    InventoryDto: {
      /** Format: int64 */
      id: number;
      /** Format: date-time */
      createDate: string;
      /** Format: date-time */
      updateDate: string;
      item: components["schemas"]["ItemDto"];
      isEquipped: boolean;
    };
// ...
    MemberMeDto: {
      /** Format: int64 */
      id: number;
      /** Format: date-time */
      createDate: string;
      /** Format: date-time */
      modifyDate: string;
      username: string;
      name: string;
      cellphoneNo: string;
      authorities: string[];
      player: components["schemas"]["PlayerDto"];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}